# Backend/src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Misinformation Diffusion Dashboard & SNA Platform API",
    "description": "Production RESTful API for Social Network Analysis, misinformation tracking, dataset management, and automated report generation.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/datasets": {
      "get": {
        "summary": "List datasets with optional search query",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Filter by filename" }
        ],
        "responses": {
          "200": { "description": "List of datasets retrieved" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create dataset record manually",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["filename"],
                "properties": {
                  "filename": { "type": "string", "example": "sample_dataset.csv" },
                  "status": { "type": "string", "example": "Uploaded" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dataset created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/datasets/upload": {
      "post": {
        "summary": "Upload CSV dataset and perform batch import",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CSV dataset uploaded and imported successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dataset uploaded successfully.",
                  "dataset": { "id": 5, "filename": "twitter_dataset.csv", "status": "Uploaded" },
                  "recordsImported": 100,
                  "summary": { "totalRows": 100, "imported": 100, "skipped": 0 }
                }
              }
            }
          },
          "400": { "description": "CSV file required or invalid file format" }
        }
      }
    },
...
    "/reports": {
      "get": {
        "summary": "Get list of generated analysis reports",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of reports" } }
      }
    },
...
  }
}